Fix area dimension precision: float columns, 4.5m seed values, no areaSizeMeters in payload
Migration 0063 changes radius_meters / length_meters / side_meters from
INTEGER to FLOAT (double precision). Integer storage forced 4m instead of
4.5m for 15 ft spells, producing a non-integer cell count on conversion
(4 / 1.5 = 2.666...). With float: 4.5 / 1.5 = 3.0 exactly.

Seed corrected:
  burning_hands: lengthMeters 4 → 4.5  (15 ft cone)
  thunderwave:   sideMeters   4 → 4.5  (15 ft cube)

Frontend:
- toNullableFloat() replaces toNullableInteger() for dimension fields so
  decimals like 4.5 are preserved in the payload
- parseOptionalPositiveFloat() used in admin helpers (allows x.5 values)
- All dimension inputs now have step={0.5}, min={0.5}

Tests:
- New: buildSpellUpdatePayload never serializes areaSizeMeters (deprecated
  field must not appear in the outbound payload)
- New: decimal dimension round-trip (4.5 → state "4.5" → payload 4.5)
- Explicit-wins-over-area_size_meters already covered in ExplicitDimensionTests

+++ apps/control-server/app/models/base_spell.py
@@ -4,7 +4,7 @@ from enum import Enum
 from typing import Optional
 from uuid import uuid4
 
-from sqlalchemy import Boolean, Column, Enum as SAEnum, ForeignKey, Integer, String, Text, UniqueConstraint
+from sqlalchemy import Boolean, Column, Enum as SAEnum, Float, ForeignKey, Integer, String, Text, UniqueConstraint
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlmodel import Field, SQLModel
 
@@ -146,17 +146,17 @@ class BaseSpell(SQLModel, table=True):
         default=None,
         sa_column=Column(Integer, nullable=True),
     )  # DEPRECATED — use radius_meters / length_meters / side_meters
-    radius_meters: Optional[int] = Field(
+    radius_meters: Optional[float] = Field(
         default=None,
-        sa_column=Column(Integer, nullable=True),
+        sa_column=Column(Float, nullable=True),
     )  # AoE radius in meters — sphere, cylinder
-    length_meters: Optional[int] = Field(
+    length_meters: Optional[float] = Field(
         default=None,
-        sa_column=Column(Integer, nullable=True),
+        sa_column=Column(Float, nullable=True),
     )  # AoE length in meters — cone, line
-    side_meters: Optional[int] = Field(
+    side_meters: Optional[float] = Field(
         default=None,
-        sa_column=Column(Integer, nullable=True),
+        sa_column=Column(Float, nullable=True),
     )  # AoE side length in meters — cube
     duration: Optional[str] = None  # editorial text
     components_json: Optional[list[str]] = Field(
